moved not TL code to src folders
